+++ R/rstudent.R
@@ -52,3 +52,30 @@ cooks.distance.deconv <- function(model, ...) {
   mse <- rss / rdf
   (r / (1 - hat))^2 * outer(hat, mse, "/") / p
 }
+
+
+## functions for fit object
+rstudent_fit <- function(fit) {
+  r <- fit$residuals
+  weights <- fit$weights
+  if (!is.null(weights)) r <- r * weights
+  hat <- fit$hat
+  rdf <- nrow(r) - ncol(fit$compensation)
+  rss <- colSums(r^2)
+  mse <- rss / nrow(r)
+  std_res <- t(t(r / sqrt(1 - hat)) / sqrt(mse))
+  # externally studentized
+  std_res * sqrt((rdf - 1) / (rdf - std_res^2))
+}
+
+cooks_distance_fit <- function(fit) {
+  r <- fit$residuals
+  weights <- fit$weights
+  if (!is.null(weights)) r <- r * weights
+  hat <- fit$hat
+  p <- ncol(fit$compensation)
+  rdf <- nrow(r) - p
+  rss <- colSums(r^2)
+  mse <- rss / rdf
+  (r / (1 - hat))^2 * outer(hat, mse, "/") / p
+}